refactor: simplify code and structure
