+++ main/app/app_hw.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <esp_err.h>
+
+esp_err_t app_hw_ensure_i2c_bus_ready(void);
