+++ main/app/app_state.c
@@ -1,0 +1,3 @@
+#include "app_state.h"
+
+app_state_t g_app = {0};
