+++ main/app/app_tasks.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void app_tasks_start(void);
+void app_tasks_stop_display(void);
+void app_display_mark_dirty(void);
